requirements.txt
numpy==2.2.6
python-dotenv==1.1.1

opencv-python==4.12.0.88



cmake==3.27.5





##Pillo==10.0.0
##scikit-learn==1.3.0


# Dlib - Especificación según la plataforma
# Para Windows con Python 3.10 (64-bit) - Usar wheel precompilado
##dlib @ https://github.com/jloh02/dlib/releases/download/v19.24.2/dlib-19.24.99-cp310-cp310-win_amd64.whl ; platform_system == "Windows" and python_version == "3.10"
face_recognition==1.3.0
# Para otras plataformas o versiones de Python
##dlib==19.24.2 ; platform_system != "Windows" or python_version != "3.10"

